Create extraction module
