+++ spec/lib/crawler/extraction/utils_spec.rb
@@ -6,15 +6,15 @@
 
 # frozen_string_literal: true
 
-RSpec.describe(Crawler::ExtractionUtils) do
+RSpec.describe(Crawler::Extraction::Utils) do
   describe '.node_descendant_text' do
     it 'should raise an error unless given a node object' do
-      expect { Crawler::ExtractionUtils.node_descendant_text('something') }.to raise_error(ArgumentError, /node-like/)
+      expect { Crawler::Extraction::Utils.node_descendant_text('something') }.to raise_error(ArgumentError, /node-like/)
     end
 
     it 'should replace break tags with spaces' do
       node = Nokogiri::HTML('<body>Hello,<br>World!')
-      expect(Crawler::ExtractionUtils.node_descendant_text(node)).to eq('Hello, World!')
+      expect(Crawler::Extraction::Utils.node_descendant_text(node)).to eq('Hello, World!')
     end
 
     context 'with uncrate.com pages' do
@@ -23,7 +23,7 @@ RSpec.describe(Crawler::ExtractionUtils) do
 
       it 'should have a reasonable performance' do
         duration = Benchmark.measure do
-          Crawler::ExtractionUtils.node_descendant_text(html)
+          Crawler::Extraction::Utils.node_descendant_text(html)
         end
 
         # It usually takes ~250 msec, used to take 180 sec before we fixed it, so let's aim for something reasonable
@@ -34,14 +34,14 @@ RSpec.describe(Crawler::ExtractionUtils) do
     context 'with ignore_tags' do
       it 'ignores <script> tags' do
         node = Nokogiri::HTML('<div><script>Script body</script><p>P body</p></div>')
-        expect(Crawler::ExtractionUtils.node_descendant_text(node)).to eq('P body')
+        expect(Crawler::Extraction::Utils.node_descendant_text(node)).to eq('P body')
       end
     end
 
     context 'without ignore_tags' do
       it 'does not ignores <script> tags' do
         node = Nokogiri::HTML('<div><script>Script body</script><p>P body</p></div>')
-        expect(Crawler::ExtractionUtils.node_descendant_text(node, [])).to eq('Script body P body')
+        expect(Crawler::Extraction::Utils.node_descendant_text(node, [])).to eq('Script body P body')
       end
     end
   end
